+++ packages/pros/examples/imu.rs
@@ -8,11 +8,11 @@ use pros::{devices::smart::SmartPort, prelude::*};
 #[derive(Default)]
 pub struct Robot;
 
-impl SyncRobot for Robot {
-    fn opcontrol(&mut self) -> pros::Result {
-        let mut imu = InertialSensor::new(unsafe { SmartPort::new(1) })?;
+impl AsyncRobot for Robot {
+    async fn opcontrol(&mut self) -> pros::Result {
+        let imu = InertialSensor::new(1)?;
 
-        imu.calibrate_blocking()?;
+        imu.calibrate().await?;
 
         loop {
             let euler = imu.euler()?;
@@ -27,4 +27,4 @@ impl SyncRobot for Robot {
     }
 }
 
-sync_robot!(Robot);
+async_robot!(Robot);