+++ packages/pros/src/devices/smart/motor.rs
@@ -12,10 +12,12 @@
 //!
 //! Example of driving a single motor with a controller:
 //! ```rust
+//! # use pros::prelude::*;
+//! let motor = Motor::new(1, BrakeMode::Brake).unwrap();
 //! let controller = Controller::Master;
 //! loop {
 //!     let output = controller.state().joysticks.left.y;
-//!     motor.set_output(output);
+//!     motor.set_output(output).ok();
 //! }
 //! ```
 
@@ -104,7 +106,7 @@ impl Motor {
         Ok(())
     }
 
-    /// Moves the motor to an absolute position, based off of the last motor zeroing.
+    /// Moves the motor to an absolute position, based off of when the motor was zeroed
     /// units for the velocity is RPM.
     pub fn set_position_absolute(
         &mut self,
@@ -181,7 +183,8 @@ impl Motor {
         }))
     }
 
-    /// Sets the current position to zero.
+    /// Sets the current encoder position to zero without moving the motor.
+    /// Analogous to taring or resetting the encoder to the current position.
     pub fn zero(&mut self) -> Result<(), MotorError> {
         unsafe {
             bail_on!(PROS_ERR, pros_sys::motor_tare_position(self.port.index()));
@@ -197,7 +200,8 @@ impl Motor {
         Ok(())
     }
 
-    /// Sets the current position to the given position.
+    /// Sets the current encoder position to the given position without moving the motor.
+    /// Analogous to taring or resetting the encoder so that the new position is equal to the given position.
     pub fn set_zero_position(&mut self, position: Position) -> Result<(), MotorError> {
         bail_on!(PROS_ERR, unsafe {
             pros_sys::motor_set_zero_position(self.port.index(), position.into_degrees())
@@ -252,6 +256,7 @@ impl SmartDevice for Motor {
 }
 
 /// Determines how a motor should act when braking.
+#[derive(Debug, Clone, Copy)]
 pub enum BrakeMode {
     /// Motor never brakes.
     None,